Version 1.4_boot1 (4/21/1994)
Bootstrap #1 for schema 1.4

This version of schema 1.4 builds using schema 1.3, and a new
schema database PI (schema_db).  Some of the files generated by
schema 1.3 need to be hand-edited.  They are here.

I'll keep this version of schema around in case I lose something
later.  Then I hope to avoid redoing the hand-edits, to get up to
the stand-alone version of schema 1.4.

(The stand-alone version will be compiled by itself).